git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@2031 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ Magick++/tests/attributes.cpp
@@ -23,8 +23,8 @@ int main( int /*argc*/, char ** argv)
 
   try {
 
-    unsigned int columns = 640;
-    unsigned int rows = 480;
+    size_t columns = 640;
+    size_t rows = 480;
     Geometry geometry(columns,rows);
     Color canvasColor( "red" );
     Image image( geometry, canvasColor);
